+++ quic/s2n-quic-transport/src/ack/ack_transmission_state.rs
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{space::rx_packet_numbers::ack_ranges::AckRanges, transmission};
+use crate::{ack::ack_ranges::AckRanges, transmission};
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum AckTransmissionState {
