+++ quic/s2n-quic-transport/src/ack/mod.rs
@@ -1,5 +1,16 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+pub use ack_manager::*;
 pub use s2n_quic_core::ack::*;
+
+mod ack_eliciting_transmission;
+mod ack_manager;
+pub(crate) mod ack_ranges;
+mod ack_transmission_state;
 pub mod interest;
+#[allow(dead_code)]
+mod pending_ack_ranges;
+
+#[cfg(test)]
+mod tests;
