--- quic/s2n-quic-transport/src/space/rx_packet_numbers/mod.rs
@@ -1,14 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-mod ack_eliciting_transmission;
-mod ack_manager;
-pub(crate) mod ack_ranges;
-mod ack_transmission_state;
-#[allow(dead_code)]
-mod pending_ack_ranges;
-
-pub use ack_manager::*;
-
-#[cfg(test)]
-mod tests;
