--- quic/s2n-quic-transport/src/space/rx_packet_numbers/tests/__fuzz__/space__rx_packet_numbers__tests__simulation/corpus.tar.gz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:44d6530d4712da762bfec0e36bbc72af905329272549a514b3aa113f36fbf6ad
-size 86274
