+++ quic/s2n-quic-transport/src/transmission/early.rs
@@ -2,10 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    contexts::WriteContext,
-    endpoint, recovery,
-    space::{rx_packet_numbers::AckManager, CryptoStream},
-    transmission,
+    ack::AckManager, contexts::WriteContext, endpoint, recovery, space::CryptoStream, transmission,
 };
 use core::ops::RangeInclusive;
 use s2n_quic_core::packet::number::PacketNumberSpace;
